FitFoodie-Backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
gunicorn==21.2.0
marshmallow==3.20.1
flask-pymongo==2.3.0
pymongo==4.6.1
dnspython==2.7.0
